Add files via upload
This code, shall determine the types of motion, based on the user's input on velocity and acceleration. In a situation, where the SI units are not used, the code will automatically convert it. The types of forces displayed are uniform, accelerated, decelerated motion, and at rest. The code is structured through selection, using the if-then, else statements, and the "elif" statements. The code consists of 3 functions in total, the function for conversion of velocity, the function for the conversion of acceleration, and the function for the types of forces based on the previous inputs on velocity and acceleration.