SG6 Code #1
This code shall ask the user to input three numbers that is accepted in the form of decimals. Based on input the code will add those numbers together, and if possible, will round off decimal digits by two digits. The built-in functions mainly consists of the input, round, and output functions.