SG6 Code #2
This code shall require the user to input the radius of a circle to calculate its area. For this one, "math" has to be imported here, since a predefined constant is required, "math.pi" specifically. In the code, the variable "area" is equal to math.pi * radius ** 2 from the mathematical formula for the area of a circle πr^2, where radius is from the user's input. The output is then printed, and is rounded off by two decimal digits.