SG 6 Activity #4
The code shall collect the user's full name and arrange it in its desired format. The code extracts the string through separating commas to separate the first, middle, and last name, and capitalize the first letter of each name. The code also checks for multiple first and last names and capitalizes them too. The middle name will be converted to middle initial. Once all these processes are done, it will arrange the substrings and display the result.